docs: Add tech-radar as a blip
Tech Radar's job is to track the technology the Open edX platform is
using. It is fitting that Tech Radar is also a blip on the Open edX Tech
Radar as it comes to fruition.

The name, ring, quadrant, and a brief description has been added. In
addition, a link to more information has been added as an anchor
element.