docs: Add tech-radar as a blip (#10)
* docs: Add tech-radar as a blip

Tech Radar's job is to track the technology the Open edX platform is
using. It is fitting that Tech Radar is also a blip on the Open edX Tech
Radar as it comes to fruition.

The name, ring, quadrant, and a brief description has been added. In
addition, a link to more information has been added as an anchor
element.

* docs: Change ring from Hold to Provisional

* docs: Change quadrant to techniques and move dirs

* refactor: remove ring dirs and update loops  (#8)

* refactor: remove ring dirs and update loops

The <blip-name>.json files have been moved from within the
`radars/primary/` directory from:
`quadrants/<quadrant-name>/rings/<ring-name>/blips/<blip-name>.json` to
`quadrants/<quadrant-name>/blips/<blip-name>.json` so that when we move
a blip from ring to ring, the git blame is no longer worthless.

In addition, the forEach loops within `scripts/export.js` have been
updated to work with the newly refactored directories. And the ring
directory has been removed from being created in `scripts/import.js`.

* fix: Sort blips by ring too

CSV is now sorted first by quadrant, then by ring, and then the blips
alphabetically.

* fix: Fix the comment  that is too long

* docs: Re-add comment about sort-values

A comment was previously removed that explained the need for the quadrant
and ring sort values. This comment has been re-added.

Co-authored-by: Carlos Muniz <cmuniz@edX-C02G655LML85.local>

* docs: Add tech-radar as a blip

Tech Radar's job is to track the technology the Open edX platform is
using. It is fitting that Tech Radar is also a blip on the Open edX Tech
Radar as it comes to fruition.

The name, ring, quadrant, and a brief description has been added. In
addition, a link to more information has been added as an anchor
element.

Co-authored-by: Carlos Muniz <cmuniz@edX-C02G655LML85.local>